+++ asyncapi/_tests/test_integration_api_http.py
@@ -2,4 +2,4 @@ import asyncapi
 
 
 def test_should_build_api_from_http_yaml():
-    assert asyncapi.api('http://localhost:9080/api.yaml')
+    assert asyncapi.build_api('http://localhost:9080/api.yaml')