+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Create subscriber cli
+
 ### 0.2.0 - 2020-07-02
 
  - Create publish and listen methods for api